rename opt module to optimize to match command
